requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
ultralytics YOLOv8